+++ components/tidb_query/src/codec/mysql/json/json_modify.rs
@@ -20,6 +20,8 @@ impl Json {
     /// Modifies a Json object by insert, replace or set.
     /// All path expressions cannot contain * or ** wildcard.
     /// If any error occurs, the input won't be changed.
+    ///
+    /// See `Modify()` in TiDB `json/binary_function.go`
     pub fn modify(
         &mut self,
         path_expr_list: &[PathExpression],
