+++ components/tidb_query/src/codec/mysql/json/json_unquote.rs
@@ -17,6 +17,8 @@ impl<'a> JsonRef<'a> {
     /// `unquote` recognizes the escape sequences shown in:
     /// https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html#
     /// json-unquote-character-escape-sequences
+    ///
+    /// See `Unquote()` in TiDB `json/binary_function.go`
     pub fn unquote(&self) -> Result<String> {
         match self.get_type() {
             JsonType::String => {
